Add functionality to sync o365 business or education drives
This patch allows to specify a driveId in the config file via the
'drive_id' parameter. If this parameter is given, the driveUrl is modified
so all requests will go to the specified driveId in stead of the users
drive.

The file README.odrive.md shows, how to get the driveId of a Sharepoint
group.